let linux generate_icon_files and generate_desktop_file to util.rs
