+++ tooling/bundler/src/bundle/linux/debian.rs
@@ -22,30 +22,20 @@
 // metadata, as well as generating the md5sums file.  Currently we do not
 // generate postinst or prerm files.
 
-use super::super::common;
+use super::{super::common, util};
 use crate::Settings;
 use anyhow::Context;
 use heck::ToKebabCase;
-use image::{self, codecs::png::PngDecoder, ImageDecoder};
 use libflate::gzip;
 use log::info;
 use walkdir::WalkDir;
 
 use std::{
   collections::BTreeSet,
-  ffi::OsStr,
   fs::{self, File},
   io::{self, Write},
   path::{Path, PathBuf},
 };
-
-#[derive(PartialEq, Eq, PartialOrd, Ord)]
-pub struct DebIcon {
-  pub width: u32,
-  pub height: u32,
-  pub is_high_density: bool,
-  pub path: PathBuf,
-}
 
 /// Bundles the project.
 /// Returns a vector of PathBuf that shows where the DEB was created.
@@ -108,7 +98,7 @@ pub fn bundle_project(settings: &Settings) -> crate::Result<Vec<PathBuf>> {
 pub fn generate_data(
   settings: &Settings,
   package_dir: &Path,
-) -> crate::Result<(PathBuf, BTreeSet<DebIcon>)> {
+) -> crate::Result<(PathBuf, BTreeSet<util::LinuxIcon>)> {
   // Generate data files.
   let data_dir = package_dir.join("data");
   let bin_dir = data_dir.join("usr/bin");
@@ -126,37 +116,10 @@ pub fn generate_data(
     .with_context(|| "Failed to copy external binaries")?;
 
   let icons =
-    generate_icon_files(settings, &data_dir).with_context(|| "Failed to create icon files")?;
-  generate_desktop_file(settings, &data_dir).with_context(|| "Failed to create desktop file")?;
+    util::generate_icon_files(settings, &data_dir).with_context(|| "Failed to create icon files")?;
+  util::generate_desktop_file(settings, &data_dir).with_context(|| "Failed to create desktop file")?;
 
   Ok((data_dir, icons))
-}
-
-/// Generate the application desktop file and store it under the `data_dir`.
-fn generate_desktop_file(settings: &Settings, data_dir: &Path) -> crate::Result<()> {
-  let bin_name = settings.main_binary_name();
-  let desktop_file_name = format!("{}.desktop", bin_name);
-  let desktop_file_path = data_dir
-    .join("usr/share/applications")
-    .join(desktop_file_name);
-  let file = &mut common::create_file(&desktop_file_path)?;
-  // For more information about the format of this file, see
-  // https://developer.gnome.org/integration-guide/stable/desktop-files.html.en
-  writeln!(file, "[Desktop Entry]")?;
-  if let Some(category) = settings.app_category() {
-    writeln!(file, "Categories={}", category.gnome_desktop_categories())?;
-  } else {
-    writeln!(file, "Categories=")?;
-  }
-  if !settings.short_description().is_empty() {
-    writeln!(file, "Comment={}", settings.short_description())?;
-  }
-  writeln!(file, "Exec={}", bin_name)?;
-  writeln!(file, "Icon={}", bin_name)?;
-  writeln!(file, "Name={}", settings.product_name())?;
-  writeln!(file, "Terminal=false")?;
-  writeln!(file, "Type=Application")?;
-  Ok(())
 }
 
 /// Generates the debian control file and stores it under the `control_dir`.
@@ -266,46 +229,6 @@ fn copy_custom_files(settings: &Settings, data_dir: &Path) -> crate::Result<()> 
     }
   }
   Ok(())
-}
-
-/// Generate the icon files and store them under the `data_dir`.
-fn generate_icon_files(settings: &Settings, data_dir: &Path) -> crate::Result<BTreeSet<DebIcon>> {
-  let base_dir = data_dir.join("usr/share/icons/hicolor");
-  let get_dest_path = |width: u32, height: u32, is_high_density: bool| {
-    base_dir.join(format!(
-      "{}x{}{}/apps/{}.png",
-      width,
-      height,
-      if is_high_density { "@2" } else { "" },
-      settings.main_binary_name()
-    ))
-  };
-  let mut icons = BTreeSet::new();
-  for icon_path in settings.icon_files() {
-    let icon_path = icon_path?;
-    if icon_path.extension() != Some(OsStr::new("png")) {
-      continue;
-    }
-    // Put file in scope so that it's closed when copying it
-    let deb_icon = {
-      let decoder = PngDecoder::new(File::open(&icon_path)?)?;
-      let width = decoder.dimensions().0;
-      let height = decoder.dimensions().1;
-      let is_high_density = common::is_retina(&icon_path);
-      let dest_path = get_dest_path(width, height, is_high_density);
-      DebIcon {
-        width,
-        height,
-        is_high_density,
-        path: dest_path,
-      }
-    };
-    if !icons.contains(&deb_icon) {
-      common::copy_file(&icon_path, &deb_icon.path)?;
-      icons.insert(deb_icon);
-    }
-  }
-  Ok(icons)
 }
 
 /// Create an empty file at the given path, creating any parent directories as
