+++ tooling/bundler/src/bundle/linux/mod.rs
@@ -6,3 +6,4 @@ pub mod appimage;
 pub mod debian;
 pub mod rpm;
 pub mod pacman;
+mod util;
